move microphones into module
