+++ crates/hulk_booster/src/main.rs
@@ -28,7 +28,6 @@ use crate::{
     hardware_interface::{BoosterHardwareInterface, Parameters as HardwareParameters},
 };
 
-mod audio_parameter_deserializers;
 mod hardware_interface;
 mod latest_receiver;
 mod microphones;
